updating to convolution module
