+++ braindecode/modules/modules.py
@@ -8,7 +8,6 @@ import numpy as np
 import torch
 import torch.nn.functional as F
 from braindecode.functional import drop_path, safe_log
-from braindecode.util import np_to_th
 from torch import Tensor, nn
 
 
@@ -60,162 +59,6 @@ class Expression(nn.Module):
         else:
             expression_str = repr(self.expression_fn)
         return self.__class__.__name__ + "(expression=%s) " % expression_str
-
-
-class SafeLog(nn.Module):
-    r"""
-    Safe logarithm activation function module.
-
-    :math:\text{SafeLog}(x) = \log\left(\max(x, \epsilon)\right)
-
-    Parameters
-    ----------
-    eps : float, optional
-        A small value to clamp the input tensor to prevent computing log(0) or log of negative numbers.
-        Default is 1e-6.
-
-    """
-
-    def __init__(self, eps: float = 1e-6):
-        super().__init__()
-        self.eps = eps
-
-    def forward(self, x) -> Tensor:
-        """
-        Forward pass of the SafeLog module.
-
-        Parameters
-        ----------
-        x : torch.Tensor
-            Input tensor.
-
-        Returns
-        -------
-        torch.Tensor
-            Output tensor after applying safe logarithm.
-        """
-        return safe_log(x=x, eps=self.eps)
-
-    def extra_repr(self) -> str:
-        eps_str = f"eps={self.eps}"
-        return eps_str
-
-
-class AvgPool2dWithConv(nn.Module):
-    """
-    Compute average pooling using a convolution, to have the dilation parameter.
-
-    Parameters
-    ----------
-    kernel_size: (int,int)
-        Size of the pooling region.
-    stride: (int,int)
-        Stride of the pooling operation.
-    dilation: int or (int,int)
-        Dilation applied to the pooling filter.
-    padding: int or (int,int)
-        Padding applied before the pooling operation.
-    """
-
-    def __init__(self, kernel_size, stride, dilation=1, padding=0):
-        super(AvgPool2dWithConv, self).__init__()
-        self.kernel_size = kernel_size
-        self.stride = stride
-        self.dilation = dilation
-        self.padding = padding
-        # don't name them "weights" to
-        # make sure these are not accidentally used by some procedure
-        # that initializes parameters or something
-        self._pool_weights = None
-
-    def forward(self, x):
-        # Create weights for the convolution on demand:
-        # size or type of x changed...
-        in_channels = x.size()[1]
-        weight_shape = (
-            in_channels,
-            1,
-            self.kernel_size[0],
-            self.kernel_size[1],
-        )
-        if self._pool_weights is None or (
-            (tuple(self._pool_weights.size()) != tuple(weight_shape))
-            or (self._pool_weights.is_cuda != x.is_cuda)
-            or (self._pool_weights.data.type() != x.data.type())
-        ):
-            n_pool = np.prod(self.kernel_size)
-            weights = np_to_th(np.ones(weight_shape, dtype=np.float32) / float(n_pool))
-            weights = weights.type_as(x)
-            if x.is_cuda:
-                weights = weights.cuda()
-            self._pool_weights = weights
-
-        pooled = F.conv2d(
-            x,
-            self._pool_weights,
-            bias=None,
-            stride=self.stride,
-            dilation=self.dilation,
-            padding=self.padding,
-            groups=in_channels,
-        )
-        return pooled
-
-
-class DepthwiseConv2d(torch.nn.Conv2d):
-    """
-    Depthwise convolution layer.
-
-    This class implements a depthwise convolution, where each input channel is
-    convolved separately with its own filter (channel multiplier), effectively
-    performing a spatial convolution independently over each channel.
-
-    Parameters
-    ----------
-    in_channels : int
-        Number of channels in the input tensor.
-    depth_multiplier : int, optional
-        Multiplier for the number of output channels. The total number of
-        output channels will be `in_channels * depth_multiplier`. Default is 2.
-    kernel_size : int or tuple, optional
-        Size of the convolutional kernel. Default is 3.
-    stride : int or tuple, optional
-        Stride of the convolution. Default is 1.
-    padding : int or tuple, optional
-        Padding added to both sides of the input. Default is 0.
-    dilation : int or tuple, optional
-        Spacing between kernel elements. Default is 1.
-    bias : bool, optional
-        If True, adds a learnable bias to the output. Default is True.
-    padding_mode : str, optional
-        Padding mode to use. Options are 'zeros', 'reflect', 'replicate', or
-        'circular'.
-        Default is 'zeros'.
-    """
-
-    def __init__(
-        self,
-        in_channels,
-        depth_multiplier=2,
-        kernel_size=3,
-        stride=1,
-        padding=0,
-        dilation=1,
-        bias=True,
-        padding_mode="zeros",
-    ):
-        out_channels = in_channels * depth_multiplier
-        super().__init__(
-            in_channels=in_channels,
-            out_channels=out_channels,
-            kernel_size=kernel_size,
-            stride=stride,
-            padding=padding,
-            dilation=dilation,
-            groups=in_channels,
-            bias=bias,
-            padding_mode=padding_mode,
-        )
 
 
 class InceptionBlock(nn.Module):
@@ -318,60 +161,6 @@ class TimeDistributed(nn.Module):
         return out.view(b, s, -1)
 
 
-class CausalConv1d(nn.Conv1d):
-    """Causal 1-dimensional convolution
-
-    Code modified from [1]_ and [2]_.
-
-    Parameters
-    ----------
-    in_channels : int
-        Input channels.
-    out_channels : int
-        Output channels (number of filters).
-    kernel_size : int
-        Kernel size.
-    dilation : int, optional
-        Dilation (number of elements to skip within kernel multiplication).
-        Default to 1.
-    **kwargs :
-        Other keyword arguments to pass to torch.nn.Conv1d, except for
-        `padding`!!
-
-    References
-    ----------
-    .. [1] https://discuss.pytorch.org/t/causal-convolution/3456/4
-    .. [2] https://gist.github.com/paultsw/7a9d6e3ce7b70e9e2c61bc9287addefc
-    """
-
-    def __init__(
-        self,
-        in_channels,
-        out_channels,
-        kernel_size,
-        dilation=1,
-        **kwargs,
-    ):
-        assert "padding" not in kwargs, (
-            "The padding parameter is controlled internally by "
-            f"{type(self).__name__} class. You should not try to override this"
-            " parameter."
-        )
-
-        super().__init__(
-            in_channels=in_channels,
-            out_channels=out_channels,
-            kernel_size=kernel_size,
-            dilation=dilation,
-            padding=(kernel_size - 1) * dilation,
-            **kwargs,
-        )
-
-    def forward(self, X):
-        out = super().forward(X)
-        return out[..., : -self.padding[0]]
-
-
 class MaxNormLinear(nn.Linear):
     """Linear layer with MaxNorm constraining on weights.
 
@@ -445,73 +234,6 @@ class LinearWithConstraint(nn.Linear):
                 self.weight.data, p=2, dim=0, maxnorm=self.max_norm
             )
         return super(LinearWithConstraint, self).forward(x)
-
-
-class CombinedConv(nn.Module):
-    """Merged convolutional layer for temporal and spatial convs in Deep4/ShallowFBCSP
-
-    Numerically equivalent to the separate sequential approach, but this should be faster.
-
-    Parameters
-    ----------
-    in_chans : int
-        Number of EEG input channels.
-    n_filters_time: int
-        Number of temporal filters.
-    filter_time_length: int
-        Length of the temporal filter.
-    n_filters_spat: int
-        Number of spatial filters.
-    bias_time: bool
-        Whether to use bias in the temporal conv
-    bias_spat: bool
-        Whether to use bias in the spatial conv
-
-    """
-
-    def __init__(
-        self,
-        in_chans,
-        n_filters_time=40,
-        n_filters_spat=40,
-        filter_time_length=25,
-        bias_time=True,
-        bias_spat=True,
-    ):
-        super().__init__()
-        self.bias_time = bias_time
-        self.bias_spat = bias_spat
-        self.conv_time = nn.Conv2d(
-            1, n_filters_time, (filter_time_length, 1), bias=bias_time, stride=1
-        )
-        self.conv_spat = nn.Conv2d(
-            n_filters_time, n_filters_spat, (1, in_chans), bias=bias_spat, stride=1
-        )
-
-    def forward(self, x):
-        # Merge time and spat weights
-        combined_weight = (
-            (self.conv_time.weight * self.conv_spat.weight.permute(1, 0, 2, 3))
-            .sum(0)
-            .unsqueeze(1)
-        )
-
-        # Calculate bias term
-        if not self.bias_spat and not self.bias_time:
-            bias = None
-        else:
-            bias = 0
-            if self.bias_time:
-                bias += (
-                    self.conv_spat.weight.squeeze()
-                    .sum(-1)
-                    .mm(self.conv_time.bias.unsqueeze(-1))
-                    .squeeze()
-                )
-            if self.bias_spat:
-                bias += self.conv_spat.bias
-
-        return F.conv2d(x, weight=combined_weight, bias=bias, stride=(1, 1))
 
 
 class MLP(nn.Sequential):
@@ -623,36 +345,6 @@ class DropPath(nn.Module):
     # Utility function to print DropPath module
     def extra_repr(self) -> str:
         return f"p={self.drop_prob}"
-
-
-class LogActivation(nn.Module):
-    """Logarithm activation function."""
-
-    def __init__(self, epsilon: float = 1e-6, *args, **kwargs):
-        """
-        Parameters
-        ----------
-        epsilon : float
-            Small float to adjust the activation.
-        """
-        super().__init__(*args, **kwargs)
-        self.epsilon = epsilon
-
-    def forward(self, x: torch.Tensor) -> torch.Tensor:
-        return torch.log(x + self.epsilon)  # Adding epsilon to prevent log(0)
-
-
-class Conv2dWithConstraint(nn.Conv2d):
-    def __init__(self, *args, max_norm=1, **kwargs):
-        self.max_norm = max_norm
-        super(Conv2dWithConstraint, self).__init__(*args, **kwargs)
-
-    def forward(self, x):
-        with torch.no_grad():
-            self.weight.data = torch.renorm(
-                self.weight.data, p=2, dim=0, maxnorm=self.max_norm
-            )
-        return super(Conv2dWithConstraint, self).forward(x)
 
 
 class FeedForwardBlock(nn.Sequential):
